Add models and overall clean up
Refactored the cli code. Improved the edit command.
