+++ cmd/openai.go
@@ -1,0 +1,5 @@
+package main
+
+import "github.com/noclue/openai/cmd/openaictl"
+
+func main() { openaictl.Run() }
